Discount Effects
Generating report including product IDs and discount effects on whether the increase in the discount rate positively impacts the number of orders for the products.
In this project, we  are expected to generate a solution using SQL with a logical approach. 
IMPORTANT: According to our approach, we can decide the percentage of positivity or negativity.